requirements.txt
python-dotenv==1.1.1
langchain==0.3.27
langchain-community==0.3.28
langchain-groq==0.3.7
streamlit==1.52.2
setuptools==80.9.0